--- a/benchmarks/brute_force_vs_bvh/brute_force_vs_bvh.hpp
+++ b/benchmarks/brute_force_vs_bvh/brute_force_vs_bvh.hpp
@@ -1,0 +1,18 @@
+/****************************************************************************
+ * Copyright (c) 2017-2022 by the ArborX authors                            *
+ * All rights reserved.                                                     *
+ *                                                                          *
+ * This file is part of the ArborX library. ArborX is                       *
+ * distributed under a BSD 3-clause license. For the licensing terms see    *
+ * the LICENSE file in the top-level directory.                             *
+ *                                                                          *
+ * SPDX-License-Identifier: BSD-3-Clause                                    *
+ ****************************************************************************/
+
+namespace ArborXBenchmark
+{
+
+template <int DIM>
+void run(int nprimitives, int nqueries, int nrepeats);
+
+}