on my laptop 
```
// THIS PR
cmake --build cmake-build-default --target ArborX_BruteForceVsBVH.exe  27.99s user 0.81s system 510% cpu 5.642 total
// ORIGIN
cmake --build cmake-build-default --target ArborX_BruteForceVsBVH.exe  16.49s user 0.48s system 99% cpu 17.097 total